Refactor and break out Tailwind colors
