+++ package.json
@@ -148,6 +148,7 @@
     "eslint-plugin-vitest": "^0.0.34",
     "esno": "^0.16.3",
     "express": "^4.18.2",
+    "fast-glob": "^3.3.1",
     "google-protobuf": "^3.17.3",
     "histoire": "^0.16.1",
     "husky": "^8.0.3",
@@ -169,11 +170,13 @@
     "stylelint": "^15.1.0",
     "stylelint-config-recommended": "^10.0.1",
     "stylelint-config-standard": "^30.0.1",
+    "svast": "^0.2.0",
+    "svast-stringify": "^0.1.0",
     "svelte": "^3.55.1",
     "svelte-check": "^3.4.3",
     "svelte-highlight": "^3.4.0",
     "svelte-loader": "^3.1.4",
-    "svelte-preprocess": "^4.10.7",
+    "svelte-preprocess": "^5.0.4",
     "svelte2tsx": "^0.5.23",
     "tailwindcss": "^3.2.4",
     "tar-fs": "^2.1.1",
